+++ RELEASE_NOTES.md
@@ -1,5 +1,14 @@
 # Release Notes
 
+## Version TBD
+_Date TBD_
+
+### Changes
+* Add surveillance tables to the database
+* Add pending surveillance tables to the database
+
+---
+
 ## Version 6.0.0
 _15 November 2016_